doc: test fixture documentation fix #12
This commit brings some precisions about test fixtures, that permit to
check the validity of the geometric feature set computation.

There is one `.png` image for each fixture, that depicts a typical
point cloud, with its corresponding eigenvectors. A description
complete the plot, by exploiting the eigenvalues/eigenvectors.

As an optional extension of this documentation effort, we could
describe the value of most important geometric features once their
computation will be OK.