featurize: introduce the '--label-scene' option
You can pass this function to extract features from a sample of a complete scene
which is 'labeled', i.e. you know the label of the sample.

The input file name should have this pattern 'location_<label>.las' for instance.

Features extraction uses the neighborhood of the complete scene to extract the
features.

related to #71