postprocess: fix the kd-tree usage
Before this commit, we used to deserialized the kd-tree computed
earlier in the pipeline. However the points are sorted when they are
read from hdf5 files. Hence the kd-tree becomes obsolete. A new
kd-tree structure must be computed with fresher data.

Moreover this commit introduce a new CLI parameter to tune this new
kd-tree: we must determine how many neighbors we want to request to
define the postprocessing step.