kmean: fix the feature updating process
The features may take the <features>_XX format, with XX the number of
neighbors with which they have been computed. Hence the feature update
does not detect them any more, as we indicate simple format in
configuration files. This commit fixes this point, by generalizing the
updating procedure (with the notable exception of z, r, g, b features,
that are still on the same format).

Moreover, the accumulation feature computing has been encapsulated in
a proper way.